+++ CHANGELOG.md
@@ -24,7 +24,7 @@ This is a complete rewrite based with a Buffer implementation. There are no majo
 
 New features:
 
-* new options from_line and to_line
+* new options info, from_line and to_line
 * trim: respect `ltrim` and `rtrim` when defined
 * delimiter: may be a Buffer
 * delimiter: handle multiple bytes/characters